Fixed 12 hours time parsing
