+++ src/ui/controls/time-range-picker/index.tsx
@@ -1,6 +1,7 @@
 import {html} from 'malevic';
 import TextBox from '../textbox';
 import {getUILanguage} from '../../../utils/locales';
+import {parseTime} from '../../../utils/time';
 
 interface TimePickerProps {
     startTime: string;
@@ -9,31 +10,6 @@ interface TimePickerProps {
 }
 
 const is12H = (new Date()).toLocaleTimeString(getUILanguage()).endsWith('M');
-
-function parseTime($time: string) {
-    const parts = $time.split(':').slice(0, 2);
-    const lowercased = $time.trim().toLowerCase();
-    const isAM = lowercased.endsWith('am') || lowercased.endsWith('a.m.');
-    const isPM = lowercased.endsWith('pm') || lowercased.endsWith('p.m.');
-
-    let hours = parts.length > 0 ? parseInt(parts[0]) : 0;
-    if (isNaN(hours) || hours > 23) {
-        hours = 0;
-    }
-    if (isAM && hours === 12) {
-        hours = 0;
-    }
-    if (isPM) {
-        hours += 12;
-    }
-
-    let minutes = parts.length > 1 ? parseInt(parts[1]) : 0;
-    if (isNaN(minutes) || minutes > 59) {
-        minutes = 0;
-    }
-
-    return [hours, minutes];
-}
 
 function toLocaleTime($time: string) {
     const [hours, minutes] = parseTime($time);
